A new freature add
A note-book created to write a python script for smart queuing system.
In this code file a Queue file has been created to deals with queue.
A personDetect class to load a model and perform an inference on the images to detect person.

More work is required to done to complete this code file according to requirments.